+++ storefront/src/app/[channel]/(main)/orders/page.tsx
@@ -3,6 +3,8 @@ import { executeGraphQL } from "@/lib/graphql";
 import { LoginForm } from "@/ui/components/LoginForm";
 import { OrderListItem } from "@/ui/components/OrderListItem";
 
+export const dynamic = "force-dynamic";
+
 export default async function OrderPage() {
 	const { me: user } = await executeGraphQL(CurrentUserOrderListDocument, {
 		cache: "no-cache",
